+++ src/rawsock-getroute.c
@@ -134,7 +134,7 @@ int rawsock_get_default_gateway(const char *ifname, unsigned *ipv4)
     err = write(fd, (char *)rtm, sizeof_buffer);
     if (err < 0 || err != sizeof_buffer) {
         perror("write(RTM_GET)");
-        printf("----%u %u\n", err, sizeof_buffer);
+        printf("----%u %u\n", err, (unsigned)sizeof_buffer);
         close(fd);
         free(rtm);
         return -1;
